# docs/openapi-cipherline-gateway.yaml
openapi: 3.0.3
info:
  title: Cipherline — API через Gateway
  description: |
    Единое описание публичных маршрутов за `API Gateway` (базовый URL, например `http://localhost:8000`).
    Внутренние `/docs` отдельных сервисов дублируют детали; этот файл — для жюри и интеграций.
  version: "1.0.0"
servers:
  - url: http://localhost:8000
    description: Локально (docker compose, gateway)
...
paths:
...
  /api/v1/progress/cipherline/state:
    get:
      tags: [progress]
      summary: Загрузить сохранённое состояние игры
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Объект state (UserState)
        "401":
          description: Нет Bearer
        "404":
          description: Сохранений ещё нет

    put:
      tags: [progress]
      summary: Сохранить полное состояние (поле login = email из токена)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Полный UserState (JSON)
      responses:
        "200":
          description: ok
        "400":
          description: Нет login в теле
        "403":
          description: login не совпадает с email в JWT
  /api/v1/progress/leaderboard:
    get:
      tags: [progress]
      summary: Публичный лидерборд (маскированные имена)
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: sort
          schema:
            type: string
            enum: [xp, accuracy, modules]
            default: xp
      responses:
        "200":
          description: Таблица лидеров и total_players
...
  /api/v1/progress/leaderboard/me:
    get:
      tags: [progress]
      summary: Ранг текущего пользователя (нужен прогресс в БД)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: rank, total_players, entry
        "401":
          description: Нет Bearer
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...